+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 2.1.6 (unreleased)
 ------------------
 
+- Fix #118: Import handler not found .
+  (`#118 <https://github.com/zopefoundation/Products.GenericSetup/issues/118>`_)
+
 - Allow passing ``quiet=True`` to ``upgradeProfile``.
   Then we do not complain when the profile is not installed or needs no upgrade.